Refactor: Change activity_id to unsignedInteger in call_statuses

+++ database/migrations/2025_09_09_000000_create_call_statuses_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     {
         Schema::create('call_statuses', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('activity_id');
+            $table->unsignedInteger('activity_id');
             $table->string('status');
             $table->text('omschrijving')->nullable();
             $table->timestamps();
